correction dans grid

+++ src/grid.py
@@ -65,7 +65,7 @@ class SudokuGrid:
         res = []
         subList = []
         #On vient d'abord récuperer les lignes correspondantes qu'on stockd dans une liste secondaire
-        for a in range(reg_row*3,reg_row*3 + 3, 1):
+        for a in range(int(reg_row*3),int(reg_row*3 + 3), 1):
             rowindice = a*9
             for pos in range(rowindice, rowindice + 9):
                 subList.append(intToStr[pos])
